workTerritory

+++ app/src/main/java/ru/practicum/android/diploma/countries/mapper/CountryMapper.kt
@@ -7,7 +7,7 @@ object CountryMapper {
 
     fun AreaEntity.toCountry(): Country {
         return Country(
-            countryId = this.areaId.toInt(),
+            countryId = this.areaId,
             countryName = this.areaName,
         )
     }
